Add support for overriding tlsuv_socket
This patch allows developers to override `tlsuv_socket`, enabling customization of protocols (e.g., SCTP or multipath TCP) or socket options. By placing the function in a separate file, it adheres to the linking rules of static libraries without requiring the `weak` annotation.
