+++ src/tlsuv.c
@@ -639,28 +639,6 @@ int tlsuv_stream_free(tlsuv_stream_t *clt) {
     return 0;
 }
 
-uv_os_sock_t tlsuv_socket(const struct addrinfo *addr, bool blocking) {
-    uv_os_sock_t sock = socket(addr->ai_family, addr->ai_socktype, addr->ai_protocol);
-
-    int on = 1;
-    int flags;
-
-#if defined(SO_NOSIGPIPE)
-    setsockopt(sock, SOL_SOCKET, SO_NOSIGPIPE, &on, sizeof(on));
-#elif defined(F_SETNOSIGPIPE)
-    fcntl(sock, F_SETNOSIGPIPE, on);
-#endif
-
-#if defined(FD_CLOEXEC)
-    flags = fcntl(sock, F_GETFD);
-    fcntl(sock, F_SETFD, flags | FD_CLOEXEC);
-#endif
-
-    tlsuv_socket_set_blocking(sock, blocking);
-
-    return sock;
-}
-
 int tlsuv_socket_set_blocking(uv_os_sock_t s, bool blocking) {
 
 #if defined(O_NONBLOCK)
